+++ workspace/TPA1/src/main/java/usuario/PatrimonioNeto.java
@@ -39,7 +39,7 @@ public class PatrimonioNeto extends Indicador{
 	
 	
 	@Override
-	public float calcular() {//metodo que tiene que estar por ser Indicador usado de manera polimorfica
+	public int calcular() {//metodo que tiene que estar por ser Indicador usado de manera polimorfica
 		return activo - pasivo;
 		
 	}
